+++ src/components/Mosaic/MosaicCanvas/useFabric.js
@@ -1,5 +1,6 @@
 import { useRef, useCallback } from 'react';
 import { fabric } from 'fabric';
+import { MOSAIC_DATA } from '../constants';
 
 // https://stackoverflow.com/questions/37565041/how-can-i-use-fabric-js-with-react
 export const useFabric = (onChange) => {
@@ -8,11 +9,13 @@ export const useFabric = (onChange) => {
     return useCallback((node) => {
         if (node) {
             fabricRef.current = new fabric.Canvas(node);
+            MOSAIC_DATA.currentRef = fabricRef.current;
             if (onChange) {
                 disposeRef.current = onChange(fabricRef.current);
             }
         }
         else if (fabricRef.current) {
+            MOSAIC_DATA.currentRef = null;
             fabricRef.current.dispose();
             if (disposeRef.current) {
                 disposeRef.current();
